Renamed Composite to Item and eliminated the Composite namespace and IComposite interface.  The interface was not necessary as nothing is deriving their own Items and this cleans up the code greatly.
Refactored the Model namespace to use the core Item instead of ModelItem.

Refactored some of the code within the platform to allow for later instantiation of the connector and passing of namespace string to use as root.  This allows the entire structure passed by Browse() to be added directly to the model with a single AddItem() method call.  The item may need to be cloned prior to doing this but so far so good.

Added code in Program to add the platform items to the model and save it, and when reloading it deletes the previously used Platform tags.