Further App namespace implementation.  Changed the name of ActionResult to OperationResult to avoid ambiguity with MVC nomenclature.  Updated platform namespace to include DirectoryExists()
Added ShortGuid() method to Utility to return the first tuple of a Guid for web requests.  Lifespan of these requests is short enough that collissions are nearly impossible and if they happen it isnt a big deal.

Added RequestInfo class to Web namespace to encapsulate some messaging around API requests