Implemented realtime logging to signalr.  Fixed a few issues with directories and other settings stored in the config file.
Corrected a few issues with logging in the App namespace, documented AddressResolver and tweaked the code to be a little cleaner.

Changed log level of individual item build notifications to Info.  This procedure is long running and I want to print something so the user knows what is going on.